Rename ThreadLocalTail for consistency
There will be upcoming changes to create safe reader abstractions,
which means that many more types will be added. The name ThreadLocalTail
does not signify if the abstraction gives you mutable or immutable
access (although it's obvious that producers should not have mutable
access to queue tails, it's important for the naming to reflect that)

As of yet, the code is pretty much the wild west of pointer
arithmetic and unsafe derefencing. Because we want to build
safe abstractions, let's start with naming.