Remove obsolete QueueReader and QueueReaderIter
Our API is designed around using impl ConsumerHandle,
i.e. referencing the ConsumerHandle trait. Because of this,
it's impossible to offer methods that return generic queue
readers, because they don't have access to the type information.
Technically they could have it, because the type info can be
passed from the ConsumerHandleImpl, but this is not possible
because the fns signature needs to match the trait exactly.