Implement test for partial writes
Also add a TODO item to the push API. (And likely
the pop API). At them moment, push is infallible
and there's no flexibility regarding partial writes/
noop selection.

Also, infallibility is pretty terrible if we want to
handle errors with tracing.

This is why I'll institute two changes:
(1) provide a fallible version, adhereing to the Rust
naming convention (something like try_push) returning
a Result. The result can be consumed by a consuming library
for tracing. So this allows us to build infallible APIs that
with side effects.

(2) Allow specification of write behavior when the queue is full.
Ideally this is an enum. Options are noop and partial write.

If we weren't writing a wait-free queue, I can imagine you could
even specify waiting behavior like set CPU to idle, or spin on the
producer capacity.